siterussia/requirements.txt
asgiref==3.6.0
django-debug-toolbar==4.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
form==0.0.1
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
